Check and clean up orphaned kex pods every time
It seems that the signal from the parent process is not transmitted
to the command(a child process) specified by the script command.
For example, if an ssh connection is killed unexpectedly it will be in that way.
(HUP signal will not be propagated to the kex process)

As a result, the kex script disappears without executing the ensure block
and the kex pod continues to loop sleep in the cluster without being deleted.

Although I tried various methods, I judged that it is difficult
to make sure that the signal is processed.
Therefore, we decided to clean up every time we run kex.